Add files via upload
1997: driveway.exe is the Win32 app. It creates driveway.txt when run.
screen_shot_driveway.png is a screen grab from 2/9/2021 of what the app looks like.
It is not interactive. Running it produces the text file, and you can inspect the main window drawing to see if you have achieved the desired results (locating the key of the basketball court between two arcs, in particular).